[Refactor] Processor 패키지 변경 및 로그 변경
